Colocando a funcionalidade de atualizar o endereço e a de poder dessacosiar o endereço do cliente caso o endereço tenha produtos associados a ele

+++ src/main/java/com/joaoandrade/pastelaria/domain/service/crud/CadastroClienteEnderecoService.java
@@ -45,6 +45,11 @@ public class CadastroClienteEnderecoService {
 	public void deletarEnderecoDoCliente(Long clienteId, Long enderecoId) {
 		Endereco endereco = buscarEnderecoDoCliente(clienteId, enderecoId);
 
+		if (!endereco.getPedidos().isEmpty() && endereco.getCliente() != null) {
+			endereco.setCliente(null);
+			return;
+		}
+
 		try {
 			repository.deleteById(endereco.getId());
 			repository.flush();
@@ -53,6 +58,11 @@ public class CadastroClienteEnderecoService {
 					String.format("Não foi possivel deletar o Endereço pois ele estar em uso no sistema!"));
 		}
 
+	}
+
+	@Transactional
+	public Endereco atualizar(Endereco endereco) {
+		return repository.save(endereco);
 	}
 
 }
